Shuffle code around to include suite as package
