+++ prow/external-plugins/verify-conformance-release/README.org
@@ -9,71 +9,89 @@ It checks that the
 
 ** Behaviour
 #+begin_src feature :tangle ./kodata/features/verify-conformance-release.feature
-Feature: Verify Conformance Product Submission
+Feature: A cool
 
   Background:
     Given a conformance product submission PR
 
-  Scenario: a test label is set
-    Given a PR without the label <label>
-    Then add the label <label> to the PR
-    Then remove the label <label> from the PR
-    Then i will fail for some reason "idk something lol"
+  # Scenario: PR title is not empty
+  #   Given a PR title
+  #   Then the PR title is not empty
+
+  # Scenario: Is true
+  #   Given a <value>
+  #   Then the value is "true"
+
+  #   Examples:
+  #     | value |
+  #     | true  |
+  #     | true  |
+  #     | false |
+
+  Scenario: submission contains all required files
+    Given a conformance product submission PR
+    Then <file> is included in its file list
+    And <file> is not empty
 
     Examples:
-      | label         |
-      | caleb-is-sorry-that-you-see-this-label |
+      | file           |
+      | "README.md"    |
+      | "PRODUCT.yaml" |
+      | "e2e.log"      |
+      | "junit_01.xml" |
 
-  # Scenario: Files must exist in correct folders
-  #   Given the files in the PR
-  #   Then file folder structure must match regex "(v1.[0-9]{2})/(.*)/.*"
-  #   # $1 is the release version of Kubernetes
-  #   # $2 is the product name
-  #   # example: v1.23/coolthing/some.file
+  Scenario: submission has files in structure of releaseversion/productname/
+    Given the files in the PR
+    Then file folder structure matches "(v1.[0-9]{2})/(.*)/.*"
+    # $1 is the release version of Kubernetes
+    # $2 is the product name
+    # example: v1.23/coolthing/some.file
 
-  # Scenario: Check for required files
-  #   Given the required <file>
-  #   Then each <file> must not be empty
+  Scenario: submission is only one product
+    Given the files in the PR
+    Then there is only one path of folders
 
-  #   Examples:
-  #     | file         |
-  #     | README.md    |
-  #     | PRODUCT.yaml |
-  #     | e2e.log      |
-  #     | junit_01.xml |
+  Scenario: the PRODUCT.yaml metadata contains required fields
+    Given a "PRODUCT.yaml" file
+    Then the yaml file "PRODUCT.yaml" contains the required and non-empty <field>
+    And if <contentType> is set to url, the content of the url in the value of <field> matches it's <dataType>
 
-  # Scenario: PRODUCT.yaml must contain required fields
-  #   Given a "PRODUCT.yaml" file
-  #   Then the yaml must contain the required and non-empty <field>
-  #   And if <type> is "url", the content of the url in the <field>'s value must match it's <dataType>
+    Examples:
+      | field               | contentType | dataType                           |
+      | "vendor"            | "info"      | "string"                           |
+      | "name"              | "info"      | "string"                           |
+      | "version"           | "info"      | "string"                           |
+      | "type"              | "info"      | "string"                           |
+      | "description"       | "info"      | "string"                           |
+      | "website_url"       | "url"       | "text/html"                        |
+      | "repo_url"          | "url"       | "text/html"                        |
+      | "documentation_url" | "url"       | "text/html"                        |
+      | "product_logo_url"  | "url"       | "image/svg application/postscript" |
 
-  #   Examples:
-  #     | field             | contentType | dataType                         |
-  #     | vendor            | info        | string                           |
-  #     | name              | info        | string                           |
-  #     | version           | info        | string                           |
-  #     | type              | info        | string                           |
-  #     | description       | info        | string                           |
-  #     | website_url       | url         | text/html                        |
-  #     | repo_url          | url         | text/html                        |
-  #     | documentation_url | url         | text/html                        |
-  #     | product_logo_url  | url         | image/svg application/postscript |
+  Scenario: title of product submission contains Kubernetes release version and product name
+    Given the title of the PR
+    Then the title of the PR matches "(.*) (v1.[0-9]{2})[ /](.*)"
+    # $1 is the string for conformance results for
+    # $2 is the version of Kubernetes
+    # $3 is the product name
+    # example: Conformance test for v1.23 Cool Engine
 
+  Scenario: the e2e.log output contains the Kubernetes release version
+    Given a "e2e.log" file
+    Then a line of the file "e2e.log" matches "^.*e2e test version: (v1.[0-9]{2}(.[0-9]{1,2})?)$"
+    # $1 is the release version of Kubernetes
+    # $2 is the (optional) point release version of Kubernetes
+    # example: Feb 25 10:20:32.383: INFO: e2e test version: v1.23.0
 
-  # Scenario: Check product name is in PR title
-  #   Given the title of the PR
-  #   Then the title of the PR must match "(.*) (v1.[0-9]{2})[ /](.*)"
-  #   # $1 is the string for conformance results for
-  #   # $2 is the version of Kubernetes
-  #   # $3 is the product name
-  #   # example: Conformance test for v1.23 Cool Engine
+  Scenario: there are labels for tests succeeding
+    Given a list of labels in the PR
+    Then the label prefixed with <label> and ending with Kubernetes release version should be present
 
-  # Scenario: Check e2e.log for Kubernetes release version
-  #   Given a "e2e.log" file
-  #   Then a line of the file "e2e.log" must match "^.*e2e test version: (v1.[0-9]{2}(.[0-9]{1,2})?)$"
-  #   # $1 is the release version of Kubernetes
-  #   # $2 is the (optional) point release version of Kubernetes
-  #   # example: Feb 25 10:20:32.383: INFO: e2e test version: v1.23.0
+    Examples:
+      | label              |
+      | "no-failed-tests-" |
+      | "tests-verified-"  |
+    # example: no-failed-tests-v1.23
 #+end_src
 
 * Build locally
